*: replace the bazelisk, disk and repo cache with sticky disks
This change teaches restoreCache to get and mount a unique stickydisk
per key and mount it at the provided path. We require a stickydisk per
path and so we hash the path and add it to the user configured "cache key".
All other semantics around versioning of the cache and respecting the base
cache key remain the same.

On post, we teach the action to unmount and commit the relevant stickydisks.
For the moment we leave the "external" cache alone. In a follow-up we can move
that to be backed by stickydisks too.

The stickydisk.js file is heavily inspired by the useblacksmith/stickydisk
implementation.