*: revert external cache to be backed by blob store
Previously, we moved the external cache to sticky disks
from blob store without fully appreciating the nuances of its
implementation. This broke some use cases of the external cache.
Since we want to be a drop in replacement, and since our colocated
cache already offers ~400MB/s download speeds, it is better to keep
the upstream behaviour intact and revert just the external cache to
use our blob store.

There is a path in the future where we can load the entire `external/`
bazel dir on a stickydisk, but this would break the use case where a
user specifies a custom `manifest` in the `external-cache` config, to
define what files are used to compute the hash for the cache key. There
are solutions to this too, but they all over-complicate the code for
not very much in the way of performance gains.